pyproject.toml
[tool.poetry]
name = "macnotesapp"
version = "0.3.0"
description = "Work with Apple MacOS Notes.app from the command line. Also includes python interface for scripting Notes.app from your own python code."
authors = ["Rhet Turnbull <rturnbull@gmail.com>"]
license = "MIT License"
readme = "README.md"
homepage = "https://github.com/RhetTbull/macnotesapp"
repository = "https://github.com/RhetTbull/macnotesapp"
keywords = ["cli", "mac", "macos"]

[tool.poetry.dependencies]
python = ">=3.8,<3.11"
py-applescript = "^1.0.3"
click = "^8.1.3"
rich = "^12.4.4"
markdown2 = "^2.4.3"
toml = "^0.10.2"
questionary = "^1.10.0"
wheel = "^0.37.1"
readability-lxml = "^0.8.1"
requests = "^2.28.1"
validators = "^0.20.0"

[tool.poetry.dev-dependencies]
pytest = "^7.1.2"
cogapp = "^3.3.0"
wheel = "^0.37.1"
build = "^0.8.0"
pyinstaller = "^5.2"

[tool.poetry.scripts]
notes = "macnotesapp.__main__:cli_main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
